+++ Echo/Changes.md
@@ -1,11 +1,19 @@
 # Utopia Echo - List of Changes
 
 ## v1.1 (in development)
+TODO: Document options support
 ### Breaking changes
+- Refactored how **Generate** requests are constructed. 
+  - The current implementation is largely based on **GenerateParams**.
 - Streaming-based requests, as well as instances of **StreamedOllamaResponseParser** 
   now require an implicit **Logger** construction parameter
   - This was added in order to manage errors within the managed pointers 
     (possibly introduced via external change event listeners)
+### New methods
+- **Prompt**
+  - Added various `.toQuery` functions
+- **Query**
+  - Added `.toRequestParams`
 ### Other changes
 - Built with Scala v2.13.14